vsphere: Improve log messages
We have some not customer-friendly error messages that simply output a response from API (terraform->govmomi) requests without any context.
The proposal of this commit is to make it easier to identify the cause of errors, providing some clues about where to start looking at.

This is part of https://issues.redhat.com/browse/RFE-1965